+++ atom/src/behaviors.h
@@ -1,5 +1,5 @@
 /*-----------------------------------------------------------------------------
-| Copyright (c) 2013-2020, Nucleic Development Team.
+| Copyright (c) 2013-2021, Nucleic Development Team.
 |
 | Distributed under the terms of the Modified BSD License.
 |
@@ -101,6 +101,7 @@ enum Mode
     List,
     Set,
     Dict,
+    NonOptional,
     Delegate,
     CallObject,
     CallObject_Object,
@@ -134,7 +135,9 @@ enum Mode
     ContainerList,
     Set,
     Dict,
+    OptionalInstance,
     Instance,
+    OptionalTyped,
     Typed,
     Subclass,
     Enum,