+++ atom/src/member.cpp
@@ -1,5 +1,5 @@
 /*-----------------------------------------------------------------------------
-| Copyright (c) 2013-2019, Nucleic Development Team.
+| Copyright (c) 2013-2021, Nucleic Development Team.
 |
 | Distributed under the terms of the Modified BSD License.
 |
@@ -859,6 +859,12 @@ Member_methods[] = {
       "Notify the static observers for the given member and atom." },
     { "tag", ( PyCFunction )Member_tag, METH_VARARGS | METH_KEYWORDS,
       "Tag the member with metatdata. " },
+#if PY_VERSION_HEX >= 0x03090000
+    // Generic aliases have been added in 3.9 and allow to index types
+    // This removes the need to quote explicit member type annotations
+    {"__class_getitem__",       (PyCFunction)Py_GenericAlias,
+    METH_O|METH_CLASS,       PyDoc_STR("See PEP 585")},
+#endif
     { 0 } // sentinel
 };