+++ atom/src/validatebehavior.cpp
@@ -1,5 +1,5 @@
 /*-----------------------------------------------------------------------------
-| Copyright (c) 2013-2019, Nucleic Development Team.
+| Copyright (c) 2013-2021, Nucleic Development Team.
 |
 | Distributed under the terms of the Modified BSD License.
 |
@@ -93,11 +93,13 @@ Member::check_context( Validate::Mode mode, PyObject* context )
             }
             break;
         }
+        case Validate::OptionalInstance:
         case Validate::Instance:
         case Validate::Subclass:
         {
             return validate_type_tuple_types( context );
         }
+        case Validate::OptionalTyped:
         case Validate::Typed:
             if( !PyType_Check( context ) )
             {
@@ -570,10 +572,8 @@ dict_handler( Member* member, CAtom* atom, PyObject* oldvalue, PyObject* newvalu
 
 
 PyObject*
-instance_handler( Member* member, CAtom* atom, PyObject* oldvalue, PyObject* newvalue )
+non_optional_instance_handler( Member* member, CAtom* atom, PyObject* oldvalue, PyObject* newvalue )
 {
-    if( newvalue == Py_None )
-        return cppy::incref( newvalue );
     int res = PyObject_IsInstance( newvalue, member->validate_context );
     if( res < 0 )
         return 0;
@@ -584,14 +584,30 @@ instance_handler( Member* member, CAtom* atom, PyObject* oldvalue, PyObject* new
 
 
 PyObject*
-typed_handler( Member* member, CAtom* atom, PyObject* oldvalue, PyObject* newvalue )
+instance_handler( Member* member, CAtom* atom, PyObject* oldvalue, PyObject* newvalue )
 {
     if( newvalue == Py_None )
         return cppy::incref( newvalue );
+    return non_optional_instance_handler( member, atom, oldvalue, newvalue );
+}
+
+
+PyObject*
+non_optional_typed_handler( Member* member, CAtom* atom, PyObject* oldvalue, PyObject* newvalue )
+{
     PyTypeObject* type = pytype_cast( member->validate_context );
     if( PyObject_TypeCheck( newvalue, type ) )
         return cppy::incref( newvalue );
     return validate_type_fail( member, atom, newvalue, type->tp_name );
+}
+
+
+PyObject*
+typed_handler( Member* member, CAtom* atom, PyObject* oldvalue, PyObject* newvalue )
+{
+    if( newvalue == Py_None )
+        return cppy::incref( newvalue );
+    return non_optional_typed_handler( member, atom, oldvalue, newvalue );
 }
 
 
@@ -810,7 +826,9 @@ handlers[] = {
     set_handler,
     dict_handler,
     instance_handler,
+    non_optional_instance_handler,
     typed_handler,
+    non_optional_typed_handler,
     subclass_handler,
     enum_handler,
     callable_handler,